+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Breaking: Save methods now return a `SaveResult` enum instead of a bool. (#105)
+- Minor: Setting managers can be configured to only save if a change has come in through its `set` function (e.g. by changing the value of a setting) with the `SaveMethod::OnlySaveIfChanged` flag. (#105)
+
 ## v0.2.2
 
 - Bugfix: Renamed `slots` in the backup options to `numSlots` to avoid conflicts with Qt (#92)